Updates import and test suite structure for all pipline submodules
