--- tests/test_pipeline/test_data_model/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for classes defined in the ``pipeline`` module."""
