--- tests/test_pipeline/test_lc_fitting/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for classes defined in the ``pipeline`` module."""
