+++ tests/test_pipeline/test_lc_simulation/testSimulateLightCurves.py
@@ -7,7 +7,7 @@ import numpy as np
 from egon.mock import MockSource, MockTarget
 
 from snat_sim.models import ObservedCadence, SNModel
-from snat_sim.pipeline import SimulateLightCurves
+from snat_sim.pipeline.lc_simultion import SimulateLightCurves
 from tests.mock import create_mock_plasticc_light_curve
